Fixes so that the application can run + renaming
- Fixes so that the application can run again
- Renaming of PartyControlExpanded to PartyControlAdvanced
