+++ LightParty/Pages/PartyMode/Advanced/ColorGradientThree.xaml.cs
@@ -2,7 +2,6 @@
 using LightParty.Services;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
@@ -19,23 +18,25 @@ using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace LightParty.Pages.PartyMode
+namespace LightParty.Pages.PartyMode.Advanced
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
-    public sealed partial class ColorGradientTwo : Page
+    public sealed partial class ColorGradientThree : Page
     {
         private dynamic popupClass;
         private dynamic onColorClass;
 
         private Color rgbStartColor;
+        private Color rgbCenterColor;
         private Color rgbStopColor;
-        private int startColorTemperature;
-        private int stopColorTemperature;
+        private int startColorTemperature = 154;
+        private int centerColorTemperature = 200;
+        private int stopColorTemperature = 500;
         private bool? rgbMode = null;
 
-        public ColorGradientTwo()
+        public ColorGradientThree()
         {
             this.InitializeComponent();
         }
@@ -51,8 +52,10 @@ namespace LightParty.Pages.PartyMode
             onColorClass = newOnColorClass;
 
             rgbStartColor = colorGradientInformation.startColor;
+            rgbCenterColor = colorGradientInformation.centerColor;
             rgbStopColor = colorGradientInformation.endColor;
             startColorTemperature = colorGradientInformation.startColorTemperature;
+            centerColorTemperature = colorGradientInformation.centerColorTemperature;
             stopColorTemperature = colorGradientInformation.endColorTemperature;
         }
 
@@ -61,53 +64,67 @@ namespace LightParty.Pages.PartyMode
             if (LightInformation.IsInRGBMode() && rgbMode != true)
             {
                 SetColor(rgbStartColor, 0);
-                SetColor(rgbStopColor, 1);
+                SetColor(rgbCenterColor, 1);
+                SetColor(rgbStopColor, 2);
 
                 rgbMode = true;
-                onColorClass.ColorGradientTwoChanged(StartColor.Color, StopColor.Color);
+                onColorClass.ColorGradientThreeChanged(StartColor.Color, CenterColor.Color, StopColor.Color);
             }
             if (!LightInformation.IsInRGBMode() && rgbMode != false)
             {
                 SetColorTemperature(startColorTemperature, 0);
-                SetColorTemperature(stopColorTemperature, 1);
+                SetColorTemperature(centerColorTemperature, 1);
+                SetColorTemperature(stopColorTemperature, 2);
 
                 rgbMode = false;
-                onColorClass.ColorTemperatureGradientTwoChanged(Convert.ToInt32(StartColorButton.Content), Convert.ToInt32(StopColorButton.Content));
+                onColorClass.ColorTemperatureGradientThreeChanged(Convert.ToInt32(StartColorButton.Content), Convert.ToInt32(CenterColorButton.Content), Convert.ToInt32(StopColorButton.Content));
             }
         }
 
         private void StartColorButton_Click(object sender, RoutedEventArgs e)
         {
             if (LightInformation.IsInRGBMode())
-                popupClass.OpenColorRGBPickerPopup<ColorGradientTwo>(StartColor.Color, this, 0);
+                popupClass.OpenColorRGBPickerPopup<ColorGradientThree>(StartColor.Color, this, 0);
             else
-                popupClass.OpenTemperatureColorPickerPopup<ColorGradientTwo>(Convert.ToInt32(StartColorButton.Content), this, 0);
+                popupClass.OpenTemperatureColorPickerPopup<ColorGradientThree>(Convert.ToInt32(StartColorButton.Content), this, 0);
+        }
+
+        private void CenterColorButton_Click(object sender, RoutedEventArgs e)
+        {
+            if (LightInformation.IsInRGBMode())
+                popupClass.OpenColorRGBPickerPopup<ColorGradientThree>(CenterColor.Color, this, 1);
+            else
+                popupClass.OpenTemperatureColorPickerPopup<ColorGradientThree>(Convert.ToInt32(CenterColorButton.Content), this, 1);
         }
 
         private void StopColorButton_Click(object sender, RoutedEventArgs e)
         {
             if (LightInformation.IsInRGBMode())
-                popupClass.OpenColorRGBPickerPopup<ColorGradientTwo>(StopColor.Color, this, 1);
+                popupClass.OpenColorRGBPickerPopup<ColorGradientThree>(StopColor.Color, this, 2);
             else
-                popupClass.OpenTemperatureColorPickerPopup<ColorGradientTwo>(Convert.ToInt32(StopColorButton.Content), this, 1);
+                popupClass.OpenTemperatureColorPickerPopup<ColorGradientThree>(Convert.ToInt32(StopColorButton.Content), this, 2);
         }
 
         public void SetColor(Color newColor, int id)
         {
-            switch(id)
+            switch (id)
             {
                 case 0:
                     StartColor.Color = newColor;
                     StartColorButton.Content = "#" + StartColor.Color.ToString().Substring(3);
                     break;
                 case 1:
+                    CenterColor.Color = newColor;
+                    CenterColorButton.Content = "#" + CenterColor.Color.ToString().Substring(3);
+                    break;
+                case 2:
                     StopColor.Color = newColor;
                     StopColorButton.Content = "#" + StopColor.Color.ToString().Substring(3);
                     break;
             }
 
             if (rgbMode == true)
-                onColorClass.ColorGradientTwoChanged(StartColor.Color, StopColor.Color);
+                onColorClass.ColorGradientThreeChanged(StartColor.Color, CenterColor.Color, StopColor.Color);
         }
 
         public void SetColorTemperature(int newColorTemperature, int id)
@@ -119,13 +136,17 @@ namespace LightParty.Pages.PartyMode
                     StartColorButton.Content = newColorTemperature.ToString();
                     break;
                 case 1:
+                    CenterColor.Color = ColorAssistant.ConvertColorTemperatureToColor(newColorTemperature);
+                    CenterColorButton.Content = newColorTemperature.ToString();
+                    break;
+                case 2:
                     StopColor.Color = ColorAssistant.ConvertColorTemperatureToColor(newColorTemperature);
                     StopColorButton.Content = newColorTemperature.ToString();
                     break;
             }
 
             if (rgbMode == false)
-                onColorClass.ColorTemperatureGradientTwoChanged(Convert.ToInt32(StartColorButton.Content), Convert.ToInt32(StopColorButton.Content));
+                onColorClass.ColorTemperatureGradientThreeChanged(Convert.ToInt32(StartColorButton.Content), Convert.ToInt32(CenterColorButton.Content), Convert.ToInt32(StopColorButton.Content));
         }
     }
 }
