+++ LightParty/Pages/PartyMode/Advanced/ColorGradientTwo.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="LightParty.Pages.PartyMode.ColorGradientTwo"
+    x:Class="LightParty.Pages.PartyMode.Advanced.ColorGradientTwo"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LightParty.Pages.PartyMode"
+    xmlns:local="using:LightParty.Pages.PartyMode.Advanced"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
