+++ LightParty/Pages/PartyMode/Advanced/MicrophoneBrightnessOptions.xaml.cs
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Navigation;
 using LightParty.Party;
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace LightParty.Pages.PartyMode
+namespace LightParty.Pages.PartyMode.Advanced
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
