+++ LightParty/Pages/PartyMode/Advanced/RandomColorOption.xaml.cs
@@ -19,14 +19,14 @@ using Windows.System;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace LightParty.Pages.PartyMode
+namespace LightParty.Pages.PartyMode.Advanced
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
     public sealed partial class RandomColorOption : Page
     {
-        PartyControl partyControl;
+        PartyControlAdvanced partyControlAdvanced;
         dynamic colorGradient;
 
         bool canSelect = false;
@@ -42,9 +42,9 @@ namespace LightParty.Pages.PartyMode
             canSelect = true;
         }
 
-        public void GiveVariables(PartyControl newPartyControl)
+        public void GiveVariables(PartyControlAdvanced newPartyControlAdvanced)
         {
-            partyControl = newPartyControl;
+            partyControlAdvanced = newPartyControlAdvanced;
         }
 
         public void LightSelectionChanged()
@@ -84,7 +84,7 @@ namespace LightParty.Pages.PartyMode
                 colorGradient = Convert.ChangeType(ColorGradientFrame.Content, ColorGradientFrame.CurrentSourcePageType);
 
                 LightProcessingColor lightProcessingColor = new LightProcessingColor();
-                colorGradient.GiveVariables<PartyControl, LightProcessingColor>(partyControl, lightProcessingColor, PartyOptions.GetColorGradientInformation());
+                colorGradient.GiveVariables<PartyControl, LightProcessingColor>(partyControlAdvanced, lightProcessingColor, PartyOptions.GetColorGradientInformation());
             }
         }
