+++ LightParty/Pages/PartyMode/PartyControl.xaml
@@ -2,21 +2,13 @@
     x:Class="LightParty.Pages.PartyMode.PartyControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LightParty.Pages.PartyMode"
+    xmlns:local="using:LightParty.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
     Loaded="Page_Loaded">
-
-    <Page.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="/Pages/Resources/SliderStyles.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Page.Resources>
 
     <Grid>
         <Grid x:Name="UserControlGrid">
@@ -34,66 +26,8 @@
             <Grid Grid.Row="2" Visibility="Collapsed">
 
             </Grid>
-
-            <Grid Grid.Row="2" Visibility="Visible">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-
-                <TextBlock Grid.Row="0" Style="{StaticResource MediumInformation}"
-                   Text="Brightness"/>
-
-                <ComboBox x:Name="BrightnessOptionComboBox" Grid.Row="1" Margin="0,5,0,0" SelectedIndex="0" Width="240"
-                      SelectionChanged="BrightnessOptionComboBox_SelectionChanged">
-                    <x:String>Change with microphone input</x:String>
-                    <x:String>Change randomly</x:String>
-                    <x:String>Do not change</x:String>
-                </ComboBox>
-                <Frame x:Name="BrightnessOptionFrame" Grid.Row="2" Margin="0,5,0,0"/>
-
-
-                <TextBlock Grid.Row="3" Style="{StaticResource MediumInformation}" Margin="0,20,0,0"
-                   Text="Color"/>
-
-                <ComboBox x:Name="ColorOptionComboBox" Grid.Row="4" Margin="0,5,0,0" SelectedIndex="0" Width="240"
-                      SelectionChanged="ColorOptionComboBox_SelectionChanged">
-                    <x:String>Change with microphone input</x:String>
-                    <x:String>Change on input difference</x:String>
-                    <x:String>Change randomly</x:String>
-                    <x:String>Do not change</x:String>
-                </ComboBox>
-                <Frame x:Name="ColorOptionFrame" Grid.Row="5" Margin="0,5,0,0"/>
-
-                
-                <TextBlock Grid.Row="6" Style="{StaticResource MediumInformation}" Margin="0,20,0,0"
-                   Text="Output"/>
-
-                <TextBlock Grid.Row="7" Style="{StaticResource Information}" Margin="0,0,0,0"
-                   Text="Brightness Output"/>
-                <Slider Style="{StaticResource OutputStyle}" x:Name="BrightnessOutput" Grid.Row="8" Width="300" HorizontalAlignment="Left" IsEnabled="False"/>
-
-                <TextBlock Grid.Row="9" Style="{StaticResource Information}" Margin="0,15,0,0"
-                   Text="Color Output"/>
-
-                <Border Grid.Row="10" Height="35" Width="35" HorizontalAlignment="Left" Margin="0,5,0,0" BorderThickness="1" CornerRadius="3" BorderBrush="{StaticResource SystemBaseHighColor}">
-                    <Rectangle>
-                        <Rectangle.Fill>
-                            <SolidColorBrush x:Name="ColorOutput" Color="Blue"/>
-                        </Rectangle.Fill>
-                    </Rectangle>
-                </Border>
-            </Grid>
         </Grid>
+        
         <Grid x:Name="PopupGrid">
             <Grid x:Name="ColorPickerPopup" HorizontalAlignment="Center" VerticalAlignment="Center" Visibility="Collapsed" Scale="0,0,0">
                 <Grid.ScaleTransition>
