+++ LightParty/Party/PartyUIUpdaterAdvanced.cs
@@ -8,13 +8,14 @@ using Windows.UI;
 using Windows.UI.Core;
 using Q42.HueApi.ColorConverters;
 using LightParty.Services;
+using LightParty.Pages.PartyMode.Advanced;
 
 namespace LightParty.Party
 {
     /// <summary>
     /// This class updates the visible UI elements of the Party Mode page.
     /// </summary>
-    class PartyUIUpdater
+    class PartyUIUpdaterAdvanced
     {
         static dynamic sliderPageOne; //Reference to the first class which contains a sound level slider.
         static dynamic sliderPageTwo; //Reference to the second class which contains a sound level slider.
@@ -32,7 +33,7 @@ namespace LightParty.Party
         private static bool useIntervalPageTwo = false; //Whether or not the second reference is used.
         private static float randomInterval; //The current random update interval.
 
-        static Pages.PartyMode.PartyControl partyControl; //Reference to the class which contains the output display
+        private static PartyControlAdvanced partyControlAdvanced; //Reference to the class which contains the output display
 
         #region Sound level slider
 
@@ -195,9 +196,9 @@ namespace LightParty.Party
         /// Sets
         /// </summary>
         /// <param name="newPartyControl"></param>
-        public static void GiveVariablesOutput(Pages.PartyMode.PartyControl newPartyControl)
+        public static void GiveVariablesOutput(PartyControlAdvanced newPartyControlAdvanced)
         {
-            partyControl = newPartyControl;
+            partyControlAdvanced = newPartyControlAdvanced;
         }
 
         /// <summary>
@@ -212,16 +213,16 @@ namespace LightParty.Party
             await Windows.ApplicationModel.Core.CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal,
             () =>
             {
-                if (partyControl != null)
+                if (partyControlAdvanced != null)
                 {
                     if (rgbColor == null && colorTemperature == null)
-                        partyControl.UpdateOutputDisplay(brightness, null);
+                        partyControlAdvanced.UpdateOutputDisplay(brightness, null);
 
                     if (rgbColor != null && colorTemperature == null)
-                        partyControl.UpdateOutputDisplay(brightness, ColorAssistant.ConvertRGBColorToColor((RGBColor)rgbColor));
+                        partyControlAdvanced.UpdateOutputDisplay(brightness, ColorAssistant.ConvertRGBColorToColor((RGBColor)rgbColor));
 
                     if (rgbColor == null && colorTemperature != null)
-                        partyControl.UpdateOutputDisplay(brightness, ColorAssistant.ConvertColorTemperatureToColor((int)colorTemperature));
+                        partyControlAdvanced.UpdateOutputDisplay(brightness, ColorAssistant.ConvertColorTemperatureToColor((int)colorTemperature));
                 }
             });
         }
